@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:loginwithlinkedin/linked_with_login/config/api_routes.dart';
 import 'package:loginwithlinkedin/linked_with_login/data_model/auth_error_response.dart';
@@ -9,7 +8,7 @@ import 'package:loginwithlinkedin/linked_with_login/data_model/profile_error.dar
 /// Created By Guru (guru@smarttersstudio.com) on 28/06/20 12:54 PM
 ///
 
-Future<LinkedInEmail> getEmailResponse({@required String accessToken}) async {
+Future<LinkedInEmail> getEmailResponse({required String? accessToken}) async {
   final response = await http.get(Uri.parse(linkedInEmailUrl),
       headers: {'Authorization': 'Bearer $accessToken'});
 
