@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:loginwithlinkedin/linked_with_login/config/api_routes.dart';
 import 'package:loginwithlinkedin/linked_with_login/data_model/auth_error_response.dart';
@@ -10,7 +9,7 @@ import 'package:loginwithlinkedin/linked_with_login/data_model/profile_response.
 ///
 
 Future<LinkedInProfile> getProfileResponse(
-    {@required String accessToken}) async {
+    {required String? accessToken}) async {
   final response = await http.get(Uri.parse(linkedInProfileUrl),
       headers: {'Authorization': 'Bearer $accessToken'});
 
